add method to serialize model data alongside its cache key
